% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utilities.R
\name{add_r_stage}
\alias{add_r_stage}
\title{Title}
\usage{
add_r_stage(stage_name, .deps = NULL, .outs = NULL, dvc_yaml_path = NULL)
}
\arguments{
\item{dvc_yaml_path}{}
}
\value{

}
\description{
Title
}
