% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utilities.R
\name{build_dvc_args}
\alias{build_dvc_args}
\title{Build DVC arguments}
\usage{
build_dvc_args(option, values)
}
\arguments{
\item{option}{command line option as a string. Example `-d`.}

\item{values}{character vector.}
}
\value{
a character.
}
\description{
Build DVC arguments
}
