% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utilities.R
\name{combine_dvc_args}
\alias{combine_dvc_args}
\title{Combine all arguments for DVC command}
\usage{
combine_dvc_args(stage_name, .deps = NULL, .outs = NULL)
}
\arguments{
\item{stage_name}{name of the stage without any white spaces. Example: `train_model`}

\item{.deps}{character vector with paths to dependencies. Optional.}

\item{.outs}{character vector with paths of outputs. Optional.}
}
\value{
a character.
}
\description{
Combine all arguments for DVC command
}
