% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dvc_project.R
\name{create_dvc_project_gui}
\alias{create_dvc_project_gui}
\title{Create a new DVC project}
\usage{
create_dvc_project_gui(path, ...)
}
\arguments{
\item{path}{user input; path to the new project}

\item{...}{for future expansion, currently unused}
}
\description{
Create a new DVC project
}
