% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{dvc_repro_upstream}
\alias{dvc_repro_upstream}
\title{Run `dvc repro` up to the currently open stage}
\usage{
dvc_repro_upstream()
}
\value{
runs `dvc repro {stage}`
}
\description{
Run `dvc repro` up to the currently open stage
}
