% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{make_read_data_function}
\alias{make_read_data_function}
\title{Create function to read data level}
\usage{
make_read_data_function(data_level, .reader = NULL)
}
\arguments{
\item{data_level}{the level of data to read from. Example `intermediate` or `raw`.}

\item{.reader}{a reader function to be used as default. Optional.}
}
\value{
a function.
}
\description{
Create function to read data level
}
\author{
João Santiago
}
