% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utilities.R
\name{prepare_new_stage_list}
\alias{prepare_new_stage_list}
\title{Prepare new stage list for DVC YAML}
\usage{
prepare_new_stage_list(new_stage_name, .deps = NULL, .outs = NULL)
}
\arguments{
\item{new_stage_name}{string name of new stage.}

\item{.deps}{character vector with paths to dependencies. Always adds the stage script as a dependency. Optional.}

\item{.outs}{character vector with paths of outputs. Optional.}
}
\value{
a list.
}
\description{
Prepare new stage list for DVC YAML
}
