% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_utilities.R
\name{read_intermediate_result}
\alias{read_intermediate_result}
\title{Read an intermediate result}
\usage{
read_intermediate_result(filename)
}
\arguments{
\item{filename}{the filename to load as a string}
}
\value{
an R object
}
\description{
Read an intermediate result
}
\author{
João Santiago
}
