% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{read_raw_data}
\alias{read_raw_data}
\title{Read raw data}
\usage{
read_raw_data(rel_path, reader = .reader, ...)
}
\arguments{
\item{rel_path}{file path relative to `data/raw`.}

\item{reader}{function used to read the serialized data.}

\item{...}{extra arguments passed to the reader.}
}
\value{
an R object.
}
\description{
Read raw data
}
\author{
João Santiago
}
