% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_utilities.R
\name{save_intermediate_result}
\alias{save_intermediate_result}
\title{Saves data to intermediate}
\usage{
save_intermediate_result(r_object, .filename = NULL)
}
\arguments{
\item{r_object}{the R object to be saved}

\item{.filename}{optional file name to pass qs::qsave. Default is df's name}
}
\value{
saves data to intermediate folder as qs file
}
\description{
Saves data to intermediate
}
\author{
João Santiago
}
