% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{stage_footer}
\alias{stage_footer}
\title{Message to finish stage}
\usage{
stage_footer(.stage = this_stage)
}
\arguments{
\item{.stage}{the name of the stage as a string. Default behavior is to search the global environment for a `this_stage` object containing the return string of `dvthis::stage_header`.}
}
\value{
prints a closing footer
}
\description{
Message to finish stage
}
