% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_utilities.R
\name{log_stage_step}
\alias{log_stage_step}
\title{Log a stage step}
\usage{
log_stage_step(msg)
}
\arguments{
\item{msg}{log message}

\item{.merchant_alias}{alias for merchant}
}
\value{
prints message to stdout
}
\description{
Log a stage step
}
\author{
João Santiago
}
