% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{read_intermediate_result}
\alias{read_intermediate_result}
\title{Read an intermediate result}
\usage{
read_intermediate_result(rel_path, reader = .reader, ...)
}
\arguments{
\item{rel_path}{file path relative to `data/intermediate`.}

\item{reader}{function used to read the serialized data. Defaults to `qs::qread`.}

\item{...}{extra arguments passed to the reader.}
}
\value{
an R object.
}
\description{
Read an intermediate result
}
\author{
João Santiago
}
