% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{save_intermediate_result}
\alias{save_intermediate_result}
\title{Saves data to intermediate}
\usage{
save_intermediate_result(
  r_object,
  rel_path = NULL,
  serializer = qs::qsave,
  ...
)
}
\arguments{
\item{r_object}{the R object to be saved.}

\item{rel_path}{optional path (relative to `data/intermediate`) to pass on to serializer. Defaults to the object's name without an extension added to the file on disk.}

\item{serializer}{function used to serialize the data to disk. Defaults to `qs::save`.}

\item{...}{extra arguments passed to the serializer.}
}
\value{
serializes an R object to the `data/intermediate` directory.
}
\description{
Saves data to intermediate
}
\author{
João Santiago
}
