% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_utilities.R
\name{stage_footer}
\alias{stage_footer}
\title{Message to finish stage}
\usage{
stage_footer(.stage = this_stage)
}
\arguments{
\item{.stage}{the name of the stage}
}
\value{
prints a closing footer
}
\description{
Message to finish stage
}
