% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_utilities.R
\name{stage_header}
\alias{stage_header}
\title{Message to separate stages}
\usage{
stage_header(this_stage, .right = NULL)
}
\arguments{
\item{this_stage}{name of the stage}

\item{.right}{additional information to show on the right side of the header. Defaults to NULL}
}
\value{
string with name of stage
}
\description{
Message to separate stages
}
\author{
João Santiago
}
